First pass at scaling down request resources.
Added expire_when_satisfied flag to request_resources

Moved AutoscalerSummary into it's own file to avoid circular reference when importing autoscaler commands from the autoscaler

Handled the flag in the autoscaler when requested resources are met.
